Proteccion de proyectos con el logado y su bearer token para cada proyecto.

+++ src/controllers/ProjectController.ts
@@ -6,6 +6,10 @@ export class ProjectController{
     
     static createProject = async (req: Request, res: Response) => {
         const project = new Project(req.body)
+
+        //Asignamos un manager
+        project.manager = req.user._id
+
         try {
             await project.save() 
             res.send('Proyecto creado correctamente')
@@ -17,7 +21,11 @@ export class ProjectController{
     static getAllProjects = async (req: Request, res: Response) => {
         
         try {
-            const projects = await Project.find({})
+            const projects = await Project.find({
+                $or: [
+                    {manager: {$in: req.user._id}}
+                ]
+            })
             res.send(projects)
         } catch (error) {
             console.log(error)
@@ -32,6 +40,10 @@ export class ProjectController{
                 const error = new Error('Proyecto no encontrado')
                 return res.status(404).json({error: error.message})
             }
+            if(project.manager.toString() !== req.user._id.toString()){
+                const error = new Error('Acción no autorizada')
+                return res.status(404).json({error: error.message})
+            }
             res.send(project)
         } catch (error) {
             console.log(error)
@@ -44,6 +56,11 @@ export class ProjectController{
             const project = await Project.findById(id)
             if(!project){
                 const error = new Error('Proyecto no encontrado')
+                return res.status(404).json({error: error.message})
+            }
+
+            if(project.manager.toString() !== req.user._id.toString()){
+                const error = new Error('Solo el manager puede hacer esta acción')
                 return res.status(404).json({error: error.message})
             }
 
@@ -66,6 +83,12 @@ export class ProjectController{
                 const errorr = new Error('Proyecto no encontrado')
                 return res.status(404).json({error: errorr.message})
             }
+
+            if(project.manager.toString() !== req.user._id.toString()){
+                const error = new Error('Solo el manager puede hacer esta acción')
+                return res.status(404).json({error: error.message})
+            }
+
             await project.deleteOne()
             res.send('Proyecto eliminado')
         } catch (error) {
